# swagger.yaml
swagger: "2.0"
info:
  description: "Weather API application that allows users to subscribe to weather updates for their city."
  version: "1.0.0"
  title: "Weather Forecast API"
host: "localhost:8080"
basePath: "/api"
...
schemes:
  - "http"
  - "https"
paths:
  /weather:
    get:
      tags:
        - "weather"
      summary: "Get current weather for a city"
      description: "Returns the current weather forecast for the specified city using WeatherAPI.com."
      operationId: "getWeather"
      parameters:
        - name: "city"
          in: "query"
          description: "City name for weather forecast"
          required: true
          type: "string"
      produces:
        - "application/json"
      responses:
        "200":
          description: "Successful operation - current weather forecast returned"
          schema:
            type: "object"
            properties:
              temperature:
                type: "number"
                description: "Current temperature"
              humidity:
                type: "number"
                description: "Current humidity percentage"
              description:
                type: "string"
                description: "Weather description"
        "400":
          description: "Invalid request"
        "404":
          description: "City not found"
...